Fixes on the way Search Events works
Search Events not searches only through calendars that the user is either "owner" or "writer". ( Editable calendars only ). In addition any events that are not editable would should on the View with a message "You do not have permission to edit this event." and all fields would be disabled for change.